+++ src/Authenticators/LDAPLoginHandler.php
@@ -6,6 +6,14 @@ use SilverStripe\Security\MemberAuthenticator\LoginHandler;
 
 class LDAPLoginHandler extends LoginHandler
 {
+    /**
+     * @var array
+     * @config
+     */
+    private static $allowed_actions = [
+        'LoginForm',
+    ];
+
     public function loginForm()
     {
         return LDAPLoginForm::create(
